Prep work for mapLegend and getColor to accept arbitrary min/max ranges separate to indicator metadata
