+++ src/app/static/src/tests/components/plots/bubble/bubblePlot.test.ts
@@ -1,8 +1,8 @@
-import {BaseWrapper, createLocalVue, shallowMount, Wrapper} from "@vue/test-utils";
+import {createLocalVue, shallowMount, Wrapper} from "@vue/test-utils";
 import BubblePlot from "../../../../app/components/plots/bubble/BubblePlot.vue";
 import {LGeoJson, LCircleMarker, LTooltip} from "vue2-leaflet";
 import {getFeatureIndicators, getIndicatorRanges, getRadius} from "../../../../app/components/plots/bubble/utils";
-import {getColor} from "../../../../app/store/filteredData/utils";
+import {getColor} from "../../../../app/components/plots/utils";
 import MapControl from "../../../../app/components/plots/MapControl.vue";
 import FilterSelect from "../../../../app/components/plots/FilterSelect.vue";
 import {FilterOption, NestedFilterOption} from "../../../../app/generated";
@@ -148,7 +148,8 @@ describe("BubblePlot component", () => {
                             <br/>Prevalence: 0.1
                             <br/>PLHIV: 10
                         </div>`);
-        let color = getColor(0.1, propsData.indicators[1]);
+        const meta = propsData.indicators[1];
+        let color = getColor(0.1, meta);
         expect(circles.at(0).props().color).toEqual(color);
         expect(circles.at(0).props().fillColor).toEqual(color);
 
@@ -159,7 +160,7 @@ describe("BubblePlot component", () => {
                             <br/>Prevalence: 0.2
                             <br/>PLHIV: 20
                         </div>`);
-        color = getColor(0.2, propsData.indicators[1]);
+        color = getColor(0.2, meta);
         expect(circles.at(1).props().color).toEqual(color);
         expect(circles.at(1).props().fillColor).toEqual(color);
