+++ src/app/static/src/tests/filteredData/utils.test.ts
@@ -1,71 +1,6 @@
-import {
-    colorFunctionFromName,
-    getColor,
-    roundToContext
-} from "../../app/store/filteredData/utils";
-import {interpolateMagma, interpolateWarm} from "d3-scale-chromatic";
-import {NestedFilterOption} from "../../app/generated";
-
+import {roundToContext} from "../../app/store/filteredData/utils";
 
 describe("FilteredData getters", () => {
-    it("colorFunctionFromName returns color function", () => {
-        const result = colorFunctionFromName("interpolateMagma");
-        expect(result).toBe(interpolateMagma);
-    });
-
-    it("colorFunctionFromName returns default color function if named function does not exist", () => {
-        const result = colorFunctionFromName("not-a-color-function");
-        expect(result).toBe(interpolateWarm);
-    });
-
-    it("getColor calculates colour string", () => {
-        const result = getColor(0.5, {
-            min: 0,
-            max: 1,
-            colour: "interpolateGreys",
-            invert_scale: false,
-            indicator: "",
-            name: "",
-            value_column: ""
-        });
-
-        expect(result).toEqual("rgb(151, 151, 151)");
-    });
-
-    it("getColor avoids dividing by zero if min equals max", () => {
-        const result = getColor(0.5, {
-            min: 0.5,
-            max: 0.5,
-            colour: "interpolateGreys",
-            invert_scale: false,
-            indicator: "",
-            name: "",
-            value_column: ""
-        });
-
-        expect(result).toEqual("rgb(255, 255, 255)");
-    });
-
-    it("getColor can invert color function", () => {
-        const metadata = {
-            min: 0,
-            max: 1,
-            colour: "interpolateGreys",
-            invert_scale: false,
-            indicator: "",
-            name: "",
-            value_column: ""
-        };
-        const result = getColor(0, metadata);
-
-        expect(result).toEqual("rgb(255, 255, 255)"); //0 = white in interpolateGreys
-
-        metadata.invert_scale = true;
-        const invertedResult = getColor(0, metadata);
-
-        expect(invertedResult).toEqual("rgb(0, 0, 0)");
-    });
-
     it ("round to context rounds values to 1 more decimal place than the context where context is integer", () => {
         expect(roundToContext(0.1234, 1)).toBe(0.1);
     });
